Implemented time validation
New object that contains start and end LocalTime fields.
New method (timeCompare) on abstract class Card that checks the current time (parsed as an argument from the NormalState) against composed object storing the time slots.
This means the Cleaner class can override the timeCompare method to validate the additional time slots of which are added as extra fields within the Cleaner class.